move style service to theme
