+++ App/Services/styles.js
@@ -2,9 +2,3 @@
 import { Dimensions } from 'react-native';
 
 export const viewportWidth = (percentageWidth: number) => Dimensions.get('window').width * (percentageWidth / 100);
-
-export const flexRowCenter = {
-  flex: 1,
-  flexDirection: 'row',
-  justifyContent: 'center',
-};
